Add files via upload
Contains MASSIVE updates to be applied on top of and integrated. Can be installed by simply asking one AI to lead the other through both the original file and the updates, and gently checking all worked. Add Enochian expansion pack this way as well, though at this point once again always wise to consult Thomalos or Max for current instructions.

Signed-off-by: AiWhispererMax <dean.esmay@gmail.com>